setup giveaways and reorganize settings

+++ starbot/bot/starbot.py
@@ -1,11 +1,12 @@
 """Starbot Main File."""
 
 # internal modules
-import yaml
+import psutil
 import datetime
 from time import time, sleep
 from logging import Formatter, INFO, StreamHandler, getLogger
-
+import importlib
+import sys
 
 # external modules
 from discord.ext.commands import Bot
@@ -25,21 +26,21 @@ class Starbot(Bot):
 
     def __init__(self, config, logger,
                  pg: Controller,
-                 guild_settings: dict):
+                 guild_settings: dict, current_giveaways: dict):
         """Initialization."""
         self.pg = pg
-        self.guild_settings = {}
-        self.start_time = current_time(True)
+        self.guild_settings = guild_settings
+        self._loaded_extensions = []
+        self.current_giveaways = current_giveaways
+        self.start_time = datetime.datetime.utcnow()
         self.logger = logger
         self.config = config
+        self.proc = psutil.Process()
         super().__init__(command_prefix=self.get_prfx)
-
 
     @classmethod
     async def get_instance(cls):
-        """
-        async method to initialize the pg_utils class
-        """
+        """Generator for db/cache."""
         # setup logger
         logger = getLogger('star-bot')
         console_handler = StreamHandler()
@@ -57,24 +58,27 @@ class Starbot(Bot):
                 logger.debug(f'Error: {e}')
                 sleep(5)
         guild_settings = await pg.get_guild_settings()
-        return cls(Config, logger, pg, guild_settings)
+        current_giveaways = await pg.get_all_giveaways(True, logger)
+        return cls(Config, logger, pg, guild_settings, current_giveaways)
 
     async def get_prfx(self, bot, message):
         try:
-            return self.guild_settings[message.guild.id]['prefix']
+            if not isinstance(message.guild, type(None)):
+                return self.guild_settings[message.guild.id]['prefix']
+            else:
+                return '!'
         except Exception as e:
             self.logger.info(f'{e}')
             return '!'
 
     async def on_ready(self):
+        print(self.guild_settings)
         try:
-            self.guild_settings = await self.pg.get_guild_settings()
+            # self.guild_settings = await self.pg.get_guild_settings()
             self.logger.info(f'\nServers: {len(self.guild_settings)}\n'
                              f'Servers: {self.guild_settings.keys()}')
         except Exception as e:
             self.logger.warning(f'Issue getting server settings: {e}')
-        if not hasattr(self, 'uptime'):
-            self.uptime = datetime.datetime.utcnow()
         self.logger.info(f'\nLogged in\n'
                          f'Name: {self.user.name}\n'
                          f'ID: {self.user.id}\n'
@@ -88,12 +92,13 @@ class Starbot(Bot):
            (int(ctx.author.id) == int(self.config.owner_id.value)):
             await self.process_commands(ctx)
         elif not await permissions.is_blacklisted(self, ctx):
+            if not isinstance(ctx.guild, type(None)):
+                if not self.guild_settings[ctx.guild.id]['invites_allowed'] and\
+                ('discord.gg' in ctx.content):
+                    ctx.delete()
+                    return
             await self.process_commands(ctx)
         else:
-            return
-        if not self.guild_settings[ctx.guild.id]['invites_allowed'] and\
-        ('discord.gg' in ctx.content):
-            ctx.delete()
             return
 
 if __name__ == "__main__":
